Every one is welcome
Every one is more then  welcome to express they opinions/suggestions. Every opinion matters.
People in need community is for every one whom believe that better life is possible. And those whom do think existing life is good We will show that it isn't truth. We here not to blame but unite and work together for a better future for all environment.